refactor: update component logic, styling, and hierarchy - extracted the side panel logic to another component in an attempt to   make it reusable (not only used for searching)
